requirements.txt
altgraph==0.17.4
bottle==0.13.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
distlib==0.3.9
filelock==3.18.0
idna==3.10
macholib==1.16.3
modulegraph==0.19.6
mysql-connector-python==9.2.0
packaging==24.2
pbr==6.1.1
platformdirs==4.3.7
proxy_tools==0.1.0
py2app==0.28.8
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyobjc-framework-Security==11.0
pyobjc-framework-WebKit==11.0
python-dateutil==2.9.0.post0
pywebview==5.4
requests==2.32.3
six==1.17.0
stevedore==5.4.1
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.30.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
zohocrmsdk7_0==5.0.0